fix(authentication): Unit test added and comments
