+++ src/Hng.Web/Controllers/AuthenticationController.cs
@@ -37,5 +37,23 @@ namespace Hng.Web.Controllers
 
             return Ok(response);
         }
+
+        [HttpPost("register")]
+        [ProducesResponseType(typeof(SignUpResponse), StatusCodes.Status201Created)]
+        [ProducesResponseType(typeof(SignUpResponse), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(SignUpResponse), StatusCodes.Status422UnprocessableEntity)]
+        [ProducesResponseType(typeof(SignUpResponse), StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult<SignUpResponse>> UserSignUp([FromBody] UserSignUpDto body)
+        {
+            var command = new UserSignUpCommand(body);
+            var response = await _mediator.Send(command);
+
+            if (response.Data == null)
+            {
+                return BadRequest(response);
+            }
+
+            return CreatedAtAction(nameof(UserSignUp), response);
+        }
     }
 }
